修改

+++ src/main/java/com/zp/integration/pojo/dataobject/UserDO.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Data
 public class UserDO implements Serializable {
 
-    private Integer id;
+    private Long id;
     private String username;
     private String password;
     private Date birthday;
@@ -22,12 +22,12 @@ public class UserDO implements Serializable {
         super();
     }
 
-    public UserDO(Integer id, String username) {
+    public UserDO(Long id, String username) {
         this.id = id;
         this.username = username;
     }
 
-    public UserDO(Integer id, String username, String password) {
+    public UserDO(Long id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
